src/third_party/blink/web_tests/wpt_internal/display-lock/css-render-subtree/render-subtree-adds-containment.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: rendersubtree adds containment</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">

<style>
.activatable {
  render-subtree: invisible;
}
.nonactivatable {
  render-subtree: invisible skip-activation;
}
</style>
<div id="container"></div>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>

<script>
function setUp() {
  container.removeAttribute("rendersubtree");
  container.style = "";
  container.classList = "";
  assert_equals(getComputedStyle(container).contain, "none");
}

test(() => {
  setUp();
  container.classList.add("nonactivatable");
  assert_equals(getComputedStyle(container).contain, "size layout style");
}, "render-subtree:'invisible skip-activation' adds contain: size layout style;");

test(() => {
  setUp();
  container.classList.add("activatable");
  assert_equals(getComputedStyle(container).contain, "size layout style");
}, "render-subtree:'invisible' adds contain: size layout style;");

test(() => {
  setUp();
  container.classList.add("activatable");
  container.style = "contain: style;";
  assert_equals(getComputedStyle(container).contain, "size layout style");
  container.style = "contain: style layout;";
  assert_equals(getComputedStyle(container).contain, "size layout style");
  container.style = "";
  assert_equals(getComputedStyle(container).contain, "size layout style");
}, "render-subtree:invisible adds contain: size layout style, can't be overridden");

test(() => {
  setUp();
  container.classList.add("activatable");
  container.style = "contain: paint;";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
  container.style = "contain: strict;";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
  container.style = "contain: content;";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
}, "rendersubtree adds contain: size layout style and keeps paint if it exists");
</script>
